database migration for new table Customer
Customers table will be replacing CustomerProfiles table User table will be long to customer.  Previously CustomerProfiles belongs to user, this prevents the implementation of nested resource that is going to be implemented to accomodate Customer sign up.  It makes more sense for Customer to has_one User and User belongs_to Customer.